+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - [User-defined Configuration Rules](TODO) allow users to dynamically change gitlint's configuration and/or the commit *before* any other rules are applied.
 - Users can now use `self.log.debug("my message")` for debugging purposes in their user-defined rules
 - Breaking: User-defined rule id's can no longer start with 'I', as those are reserved for built-in gitlint ignore rules.
+- **New Rule**: `ignore-body-lines` allows users to
+[ignore parts of a commit](http://jorisroovers.github.io/gitlint/#ignoring-commits) by matching a regex against
+the lines in a commit message body. ([#126](https://github.com/jorisroovers/gitlint/issues/126)).
 
 ## v0.13.1 (2020-02-26)